Adding login functionality and assigning roles to the users

+++ store_app/views/MovieCRUD.py
@@ -1,12 +1,14 @@
-from flask import make_response, jsonify, request
+from flask import make_response, jsonify, request, g
 from business import ConnectExternalAPI, MovieCRUD
 from app import app
 
 
-@app.route('/movies', methods=['POST', 'DELETE'])
+@app.route('/movie', methods=['POST', 'DELETE'])
 def add_or_remove_movie():
     if request.method == "POST":
         try:
+            if g.user.role != 'admin':
+                raise ValueError('User must be an admin to perform this function.')
             imdb_id = request.json['imdb_id']
             quantity = request.json['quantity']
             price = request.json['price']
@@ -20,6 +22,8 @@ def add_or_remove_movie():
 
     if request.method == 'DELETE':
         try:
+            if g.user.role != 'admin':
+                raise ValueError('User must be an admin to perform this function.')
             imdb_id = request.json['imdb_id']
             MovieCRUD.delete_movie(imdb_id)
             msg = "Record successfully deleted"
@@ -32,6 +36,8 @@ def add_or_remove_movie():
 @app.route('/movies/<string:imdb_id>/quantity', methods=['PATCH'])
 def set_quantity(imdb_id):
     try:
+        if g.user.role != 'admin':
+            raise ValueError('User must be an admin to perform this function.')
         new_quantity = request.json['quantity']
         MovieCRUD.set_quantity(imdb_id, new_quantity)
         msg = "Record successfully updated"
@@ -44,6 +50,8 @@ def set_quantity(imdb_id):
 @app.route('/movies/<string:imdb_id>/price', methods=['PATCH'])
 def set_price(imdb_id):
     try:
+        if g.user.role != 'admin':
+            raise ValueError('User must be an admin to perform this function.')
         new_price = request.json['price']
         MovieCRUD.set_price(imdb_id, new_price)
         msg = "Record successfully updated"
@@ -56,6 +64,8 @@ def set_price(imdb_id):
 @app.route('/movies/<string:imdb_id>/year', methods=['PATCH'])
 def set_year(imdb_id):
     try:
+        if g.user.role != 'admin':
+            raise ValueError('User must be an admin to perform this function.')
         new_year = request.json['year']
         MovieCRUD.set_year(imdb_id, new_year)
         msg = "Record successfully updated"
